+++ lib/src/stream_with_value_builder.dart
@@ -32,15 +32,18 @@ class _StreamBuilderWithValueState<T> extends State<StreamBuilderWithValue<T>> {
 }
 
 typedef DataTrigger<T> = void Function(T newValue);
-typedef DataBuilder<T> = Widget Function(BuildContext context, T value);
+typedef DataBuilder<T extends Object> = Widget Function(
+    BuildContext context, T value);
 typedef NullValueBuilder = Widget Function(BuildContext context);
 
 Widget _noValueBuilder(BuildContext context) =>
     const Center(child: CircularProgressIndicator());
 
-class DataStreamWithValueBuilder<T> extends StatefulWidget {
-  /// Source [StreamWithValue].
-  final StreamWithValue<T> streamWithValue;
+class DataStreamWithValueBuilder<T extends Object> extends StatefulWidget {
+  /// Source [StreamWithValue]. We accept a nullable type here to be able to
+  /// filter out nulls without requiring the user to pass nullable type as [T]
+  /// and then using exclamation mark operator in the builder body.
+  final StreamWithValue<T?> streamWithValue;
 
   /// Builds a child widget. Never gets `null` as a value.
   final DataBuilder<T> builder;
@@ -51,7 +54,7 @@ class DataStreamWithValueBuilder<T> extends StatefulWidget {
 
   /// Called for every change to the data. May be called with `null` if [T] is
   /// a nullable type.
-  final DataTrigger<T>? onData;
+  final DataTrigger<T?>? onData;
 
   /// Called when errors happen with the Stream
   final void Function(dynamic e, StackTrace stackTrace)? onError;
@@ -62,52 +65,35 @@ class DataStreamWithValueBuilder<T> extends StatefulWidget {
     this.onData,
     this.nullValueBuilder = _noValueBuilder,
     this.onError,
-  }) : super(key: ValueKey(streamWithValue.updates));
+    Key? key,
+  }) : super(key: key);
 
   @override
   _DataStreamWithValueBuilderState<T> createState() =>
       _DataStreamWithValueBuilderState<T>();
 }
 
-class _DataStreamWithValueBuilderState<T>
+class _DataStreamWithValueBuilderState<T extends Object>
     extends State<DataStreamWithValueBuilder<T>> {
-  late StreamSubscription<T> _streamSubscription;
+  late StreamSubscription<T?> _streamSubscription;
   T? _currentValue;
 
   @override
   void initState() {
     super.initState();
+    _processCurrentValueAndSubscribeToStream();
+  }
 
-    if (widget.streamWithValue.loaded) {
-      final value = widget.streamWithValue.value;
-      _currentValue = value;
-      if (widget.onData != null) {
-        scheduleMicrotask(() => widget.onData!(value));
-      }
+  @override
+  void didUpdateWidget(DataStreamWithValueBuilder<T> oldWidget) {
+    super.didUpdateWidget(oldWidget);
+
+    if (oldWidget.streamWithValue != widget.streamWithValue) {
+      _streamSubscription.cancel();
+      _processCurrentValueAndSubscribeToStream();
+    } else if (oldWidget.onError != widget.onError) {
+      _streamSubscription.onError(widget.onError);
     }
-
-    _streamSubscription = widget.streamWithValue.updates.listen(
-      (event) {
-        if (mounted) {
-          if (widget.onData != null) {
-            widget.onData!(event);
-          }
-          setState(() {
-            _currentValue = event;
-          });
-        }
-      },
-      onDone: () {
-        if (mounted) {
-          Navigator.of(context).pop();
-        }
-      },
-      onError: (
-        dynamic e,
-        StackTrace stackTrace,
-      ) =>
-          widget.onError?.call(e, stackTrace),
-    );
   }
 
   @override
@@ -122,4 +108,35 @@ class _DataStreamWithValueBuilderState<T>
           ? widget.nullValueBuilder(context)
           : widget.builder(context, _currentValue!)
       : _noValueBuilder(context);
+
+  void _processCurrentValueAndSubscribeToStream() {
+    if (widget.streamWithValue.loaded) {
+      final value = widget.streamWithValue.value;
+      _currentValue = value;
+      if (widget.onData != null) {
+        // Schedule a microtask in case onData calls for context, which is not
+        // allowed in e.g. initState.
+        scheduleMicrotask(() => widget.onData!(value));
+      }
+    } else {
+      _currentValue = null;
+    }
+
+    _streamSubscription = widget.streamWithValue.updates.listen(
+      (event) {
+        if (mounted) {
+          widget.onData?.call(event);
+          setState(() {
+            _currentValue = event;
+          });
+        }
+      },
+      onDone: () {
+        if (mounted) {
+          Navigator.of(context).pop();
+        }
+      },
+      onError: widget.onError,
+    );
+  }
 }